+++ src/Fractal/ReleaseTransformer.php
@@ -7,9 +7,17 @@ use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
 
 class ReleaseTransformer extends TransformerAbstract
 {
-
     /** @var UrlGeneratorInterface $urlGenerator */
     protected $urlGenerator;
+
+    /** @var array $installerTypes */
+    protected $installerTypes;
+
+    public function __construct(UrlGeneratorInterface $urlGenerator, array $installerTypes = [])
+    {
+        $this->urlGenerator   = $urlGenerator;
+        $this->installerTypes = $installerTypes;
+    }
 
     /**
      * @param Release $release
@@ -31,13 +39,13 @@ class ReleaseTransformer extends TransformerAbstract
             ],
         ];
 
-        if ('library' != $release->type) {
+        if ('library' !== $release->type) {
             $package['type'] = $release->type;
         }
 
-        if (in_array($release->type, [ 'wordpress-plugin', 'wordpress-theme', 'wordpress-muplugin' ])) {
+        if (in_array($release->type, $this->installerTypes, true)) {
             $package['require'] = [
-                'composer/installers' => '~1.0'
+                'composer/installers' => '^1.5',
             ];
         }
 
@@ -46,6 +54,8 @@ class ReleaseTransformer extends TransformerAbstract
 
     /**
      * Set UrlGenerator to use for file links in output.
+     *
+     * @deprecated 0.3:1.0 Deprecated in favor of constructor.
      *
      * @param UrlGeneratorInterface $urlGenerator
      */