+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - command line boilerplate to home page.
 - site favicon, image from [Bytesize](https://danklammer.com/bytesize-icons/).
 - default error log configuration detection from PHP settings.
+- support for all `composer/installer` types (^1.5).
 
 ### Changed
 - home page markup to Bootstrap 4.